Needed to fix this... keep an eye on 64-bit build.

+++ src/rustbox-pure/console/mod.rs
@@ -72,7 +72,7 @@ fn translate_key_code(raw_event: KEY_EVENT_RECORD) -> Option<Key> {
 	if raw_event.bKeyDown == 1 {
 		// The API seems to only handle key-down events.
 
-		match(raw_event.wVirtualKeyCode as u64) {
+		match(raw_event.wVirtualKeyCode as u32) {
 			VK_TAB => Some(Key::Tab),
 			VK_RETURN => Some(Key::Enter),
 			VK_ESCAPE => Some(Key::Esc),
